+ ABB IR8700 Model
adding the first robot model.
This is a .dxf of the ABB IR8700-800 350 that I got from their website.
I converted it to an OBJ in FreeCAD and brought it into blender.
So far, the origins (pivots) have been properly set on the geo.
Next step is to add the armature and start looking at constraints.