+formatted position data for all 6 axis as geo node trees and rotating cubes.
Position data for all 6 axis is formatted properly based on the 8700 user manual. (Added a screenshot of the axis orientation for reference.)

Made the choice to take the geometry node trees off the object with the armature animation expression for 2 reasons.

1. The animated text was not readable.

2. Having the 6 animated cubes is super useful to troubleshoot isolated axis orientation issues.

Next step is to get the kinematic data calculated and then graphed.