Path Animation Scene Updates
Small Tweaks to the path animation test scene for usability.

Biggest Change: Animation for the path constrained object now lives on the track object.

Go to Data -> Path Animation and ensure frames is set to 100 (this is really the total percent value.) Then use Evaluation Time as your animation curve. Key from 0 to 100.000